Refactoring
-combined into one controllers AccountsController and AccountController;
-combined feign clients;
-added @ControllerAdvise class for handling exceptions;
-some code optimizations.
