+++ pfs-accounting/src/main/java/ru/gb/dev/spring/pfs/accounting/exception/EntityNotFoundException.java
@@ -1,0 +1,11 @@
+package ru.gb.dev.spring.pfs.accounting.exception;
+
+public class EntityNotFoundException extends RuntimeException {
+
+    private static final long serialVersionUID = 386726542537970495L;
+
+    public EntityNotFoundException(final String message) {
+        super(message);
+    }
+
+}
