+++ pfs-accounting/src/main/java/ru/gb/dev/spring/pfs/accounting/exception/ErrorDto.java
@@ -1,0 +1,20 @@
+package ru.gb.dev.spring.pfs.accounting.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+/**
+ * DTO class for errors
+ * @author V. Zinchenko
+ */
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class ErrorDto {
+    private String code;
+    private String message;
+}
