+++ pfs-accounting/src/main/java/ru/gb/dev/spring/pfs/accounting/feign/OperationsClient.java
@@ -1,9 +1,11 @@
-package ru.gb.dev.spring.pfs.accounting.controller.client;
+package ru.gb.dev.spring.pfs.accounting.feign;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import ru.gb.dev.spring.pfs.accounting.model.dto.OperationDto;
+import ru.gb.dev.spring.pfs.accounting.model.dto.util.ResultDto;
 
 import java.util.List;
 
@@ -13,7 +15,13 @@ import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8_VALUE;
 @FeignClient(name = "operation")
 public interface OperationsClient {
 
-    @GetMapping(value = "/operations",  produces = APPLICATION_JSON_UTF8_VALUE)
+    @GetMapping(value = "/ping",  produces = APPLICATION_JSON_UTF8_VALUE)
+    ResultDto ping();
+
+    @GetMapping(value = "{id}",  produces = APPLICATION_JSON_UTF8_VALUE)
+    OperationDto getOperation(@PathVariable("id") final String operationId);
+    
+    @GetMapping(produces = APPLICATION_JSON_UTF8_VALUE)
     List<OperationDto> getOperations();
 
 }
