+++ pfs-accounting/src/main/java/ru/gb/dev/spring/pfs/accounting/feign/UsersClient.java
@@ -1,7 +1,8 @@
-package ru.gb.dev.spring.pfs.accounting.controller.client;
+package ru.gb.dev.spring.pfs.accounting.feign;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import ru.gb.dev.spring.pfs.accounting.model.dto.UserDto;
 
@@ -13,7 +14,10 @@ import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8_VALUE;
 @FeignClient(name = "user")
 public interface UsersClient {
 
-    @GetMapping(value = "/users", produces = APPLICATION_JSON_UTF8_VALUE)
+    @GetMapping(value = "{id}", produces = APPLICATION_JSON_UTF8_VALUE)
+    UserDto getOperation(@PathVariable("id") final String userId);
+
+    @GetMapping(produces = APPLICATION_JSON_UTF8_VALUE)
     List<UserDto> getUsers();
 
 }
