--- pfs-accounting/src/main/java/ru/gb/dev/spring/pfs/accounting/util/NotFoundException.java
@@ -1,11 +1,0 @@
-package ru.gb.dev.spring.pfs.accounting.util;
-
-public class NotFoundException extends RuntimeException {
-
-    private static final long serialVersionUID = 386726542537970495L;
-
-    public NotFoundException(final String message) {
-        super(message);
-    }
-
-}
